+++ example_peps-cfg2/example_BiocProject/readBedFiles_resize.R
@@ -1,0 +1,14 @@
+readBedFiles_resize = function(project, resize.width) {
+  cwd = getwd()
+  paths = pepr::sampleTable(project)$file_path
+  sampleNames = pepr::sampleTable(project)$sample_name
+  setwd(dirname(project@file))
+  result = lapply(paths, function(x){
+    df = read.table(x)
+    colnames(df) = c('chr', 'start', 'end')
+    gr = GenomicRanges::resize(GenomicRanges::GRanges(df), width=resize.width)
+  })
+  setwd(cwd)
+  names(result) = sampleNames
+  return(GenomicRanges::GRangesList(result))
+}