+++ peppy/__init__.py
@@ -10,25 +10,10 @@ local level, but this will at least provide a foundation.
 from ._version import __version__
 from .const import *
 from .exceptions import *
-from .project import Project, ProjectContext
-from .sample import Sample, Subsample
-from .snake_project import *
-from .utils import fetch_samples, grab_project_data, CommandChecker
-#from logmuse import init_logger
+from .project import Project
+from .sample import Sample
 
-_EXPORT_FROM_UTILS = [fetch_samples.__name__, grab_project_data.__name__,
-                      CommandChecker.__name__]
-
-__classes__ = ["Project", "Sample", "SnakeProject"]
-__all__ = __classes__ + ["PeppyError"] + _EXPORT_FROM_UTILS
-
-# TODO: remove
-COMPUTE_SETTINGS_VARNAME = "DIVCFG"
+__classes__ = ["Project", "Sample"]
+__all__ = __classes__ + ["PeppyError"]
 
 LOGGING_LEVEL = "INFO"
-
-# Ensure that we have a handler and don't get a logging exception.
-#_LOGGER = logging.getLogger(__name__)
-#if not logging.getLogger().handlers:
-#    _LOGGER.addHandler(logging.NullHandler())
-#_LOGGER = init_logger(name="peppy", level=LOGGING_LEVEL)