+++ update_examples.sh
@@ -1,6 +1,0 @@
-wget https://github.com/pepkit/example_peps/archive/master.zip
-mv master.zip examples
-cd examples
-rm -rf example_peps-master 
-unzip master.zip
-rm master.zip