+++ docs/changelog.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [0.30.0] - 2020-05-26
+
+**This version introduced backwards-incompatible changes.**  
+
+### Added
+- attribute duplication functionality
+- config importing functionality
+- attribute removal functionality
+- possibility to define multi-attribute rules in attribute implication
+  
+### Changed
+- Project configuration file to follow [PEP2.0.0](http://pep.databio.org/en/2.0.0/specification/) specification. Browse the specification for changes related to config format
+- Do not require `sample_name` attribute in the sample table
+
 ## [0.22.3] - 2019-12-13
 ### Changed
 - Remove `is_command_callable` from `utils` module; instead, refer to [`ubiquerg`](https://pypi.org/project/ubiquerg/).